projeto/requirements.txt
faiss-cpu==1.11.0
llama-index==0.11.20
llama-index-core==0.11.23
llama-index-vector-stores-faiss==0.2.1
llama-index-embeddings-huggingface==0.3.1
llama-index-llms-groq==0.2.0
crewai==0.130.0
streamlit==1.44.1
python-dotenv==1.0.1